refactor: switch to Flask Blueprints for better project structure
I changed the initial idea of the implementation and moved to Flask Blueprints
to improve the structure, organization, and scalability of the project.

In the process, I entered into an endless loop of errors, so I had to stop
committing until I resolved them.

This commit represents one of the final stages of the project and differs
significantly from the initial commits.